requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Explainability
grad-cam>=1.5.0

# Experiment tracking
mlflow>=2.8.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
